get comment from server

+++ InitDb.sql
@@ -3,12 +3,12 @@ CREATE DATABASE teach;
 USE teach;
 
 CREATE TABLE teach.user (
-  id         INT UNSIGNED PRIMARY KEY NOT NULL AUTO_INCREMENT,
-  login_name VARCHAR(30)              NOT NULL,
-  password   CHAR(16)                 NOT NULL,
-  nick_name  NVARCHAR(20)             NOT NULL,
+  id         BIGINT UNSIGNED PRIMARY KEY NOT NULL AUTO_INCREMENT,
+  login_name VARCHAR(30)                 NOT NULL,
+  password   CHAR(16)                    NOT NULL,
+  nick_name  NVARCHAR(20)                NOT NULL,
   protrait   VARCHAR(20) COMMENT '用户头像文件名',
-  phone      VARCHAR(20)              NOT NULL,
+  phone      VARCHAR(20)                 NOT NULL,
   token      CHAR(20),
   remaining  DECIMAL COMMENT '余额'
 );
@@ -19,7 +19,7 @@ CREATE UNIQUE INDEX user_name_phone_uindex
 
 
 CREATE TABLE teach.teacher (
-  id           INT UNSIGNED PRIMARY KEY NOT NULL AUTO_INCREMENT,
+  id           BIGINT UNSIGNED PRIMARY KEY NOT NULL AUTO_INCREMENT,
   name         NVARCHAR(30),
   subject      NVARCHAR(30) COMMENT '涉及学科',
   graduate     NVARCHAR(30) COMMENT '毕业院校',
@@ -32,37 +32,37 @@ ALTER TABLE teach.teacher
 
 
 CREATE TABLE teach.course (
-  id           INT UNSIGNED PRIMARY KEY NOT NULL AUTO_INCREMENT,
-  name         NVARCHAR(40)             NOT NULL COMMENT '课程名称',
+  id           BIGINT UNSIGNED PRIMARY KEY NOT NULL AUTO_INCREMENT,
+  name         NVARCHAR(40)                NOT NULL COMMENT '课程名称',
   introdcution NVARCHAR(100) COMMENT '课程描述',
-  teacher_id    INT UNSIGNED             NOT NULL COMMENT '教师id',
-  video        VARCHAR(30)              NOT NULL COMMENT '课程视频文件',
-  price        DECIMAL                  NOT NULL,
-  image        VARCHAR(30)              NOT NULL
+  teacher_id   BIGINT UNSIGNED             NOT NULL COMMENT '教师id',
+  video        VARCHAR(30)                 NOT NULL COMMENT '课程视频文件',
+  price        DECIMAL                     NOT NULL,
+  image        VARCHAR(30)                 NOT NULL
 );
 
 CREATE TABLE teach.join_course_user
 (
-  id        INT UNSIGNED PRIMARY KEY NOT NULL AUTO_INCREMENT,
-  user_id   INT UNSIGNED             NOT NULL,
-  course_id INT UNSIGNED             NOT NULL
+  id        BIGINT UNSIGNED PRIMARY KEY NOT NULL AUTO_INCREMENT,
+  user_id   BIGINT UNSIGNED             NOT NULL,
+  course_id BIGINT UNSIGNED             NOT NULL
 );
 
 CREATE TABLE teach.comment
 (
-  id        INT UNSIGNED PRIMARY KEY  NOT NULL  AUTO_INCREMENT,
-  user_id   INT UNSIGNED              NOT NULL,
-  course_id INT UNSIGNED              NOT NULL,
+  id        BIGINT UNSIGNED PRIMARY KEY  NOT NULL  AUTO_INCREMENT,
+  user_id   BIGINT UNSIGNED              NOT NULL,
+  course_id BIGINT UNSIGNED              NOT NULL,
   content   NVARCHAR(100),
-  tiemstamp TIMESTAMP                 NOT NULL
+  tiemstamp TIMESTAMP                    NOT NULL
 );
 
 
 CREATE TABLE teach.verification
 (
-  id      INT UNSIGNED PRIMARY KEY  NOT NULL  AUTO_INCREMENT,
-  code    VARCHAR(5)                NOT NULL,
-  picture VARCHAR(30)               NOT NULL
+  id      BIGINT UNSIGNED PRIMARY KEY  NOT NULL  AUTO_INCREMENT,
+  code    VARCHAR(5)                   NOT NULL,
+  picture VARCHAR(30)                  NOT NULL
 );
